+++ CHANGELOG.md
@@ -7,11 +7,15 @@ described with adjustments and template in [TSWiki](https://tswiki.corp.saab.se/
 
 ## 4.7
 
+### Added
+- Support for configure foliage occlusion 
+
 ### Changed
 - Updated to GizmoSDK 2.12.326.1
 
 ### Fixed
 - Null reference exception when unloading a map containing asset instances #309075
+- Foliage Occlusion on Embedded Linux #407581
 
 ## 4.6